fix: voice connection issues
we are currently experiencing some issues with voice calls where the bot randomly gets disconnected
with close code 4006. this is apparently an upstream issue: http://github.com/Rapptz/discord.py/pull/10210

i don't really want to move straight to the master branch even if this PR gets merged, so
i'm cutting out this specific patch to fix the library while we wait for upstream to make a new release.